Updated to prevent division by zero cases.
I updated the algorithm in order to deal with coefficients of zero. For the current row, if the column of interest stores a value of zero, the algorithm will try to swap the place of this row with another row below it that has a non-zero value. If no such row can be found, this row is skipped.